Added separate page for editing profile page
