+++ src/main/webapp/WEB-INF/web.xml
@@ -77,6 +77,16 @@
   </servlet-mapping>
 
   <servlet>
+    <servlet-name>ProfileEditServlet</servlet-name>
+    <servlet-class>codeu.controller.ProfileEditServlet</servlet-class>
+  </servlet>
+
+  <servlet-mapping>
+    <servlet-name>ProfileEditServlet</servlet-name>
+    <url-pattern>/profileEdit</url-pattern>
+  </servlet-mapping>
+
+  <servlet>
     <servlet-name>RegisterServlet</servlet-name>
     <servlet-class>codeu.controller.RegisterServlet</servlet-class>
   </servlet>
