+++ src/main/webapp/css/main.css
@@ -135,3 +135,9 @@ input {
 button {
   font-size: 18px;
 }
+
+#moments{
+  list-style-type:none;
+  border:2px;
+  padding: 0;
+}
